+++ pycloudflare/__init__.py
@@ -1,4 +1,4 @@
 """Python client for CloudFlare."""
 
-__version__ = '0.4.5'
+__version__ = '1.0.0'
 __url__ = 'https://github.com/yola/pycloudflare'