+++ pycloudflare/services.py
@@ -122,14 +122,18 @@ class CloudFlareService(HTTPServiceClient):
     def delete_page_rule(self, zone_id, rule_id):
         return self.delete('zones/%s/pagerules/%s' % (zone_id, rule_id))
 
-    def purge_cache(self, zone_id, files=None, tags=None):
+    def purge_cache(self, zone_id, files=None, tags=None, hosts=None):
         data = {}
         if files:
             data['files'] = files
         if tags:
             data['tags'] = tags
-        if files is None and tags is None:
+        if hosts:
+            data['hosts'] = hosts
+
+        if not data:
             data['purge_everything'] = True
+
         return self.delete('zones/%s/purge_cache' % zone_id, json=data)
 
     def get_ssl_universal_settings(self, zone_id):